--- client/src/components/NavBarTest.jsx
@@ -1,100 +1,0 @@
-import React from "react";
-import { GrTrophy } from "react-icons/gr";
-
-class NavBarTest extends React.Component {
-  constructor(props) {
-    super(props);
-
-    this.clickHandler = this.clickHandler.bind(this);
-  }
-
-  clickHandler({ target }) {
-    if (target.innerText === "About") {
-      this.props.handleClick();
-    }
-  }
-  render() {
-    return (
-      <nav className="navbar navbar-expand-lg navbar-light bg-light">
-        <a className="navbar-brand" href="#">
-          <GrTrophy />
-        </a>
-        {/* <a style={{ fontWeight: "bold", paddingRight: "3px" }}>MS</a> */}
-        <button
-          className="navbar-toggler"
-          type="button"
-          data-toggle="collapse"
-          data-target="#navbarSupportedContent"
-          aria-controls="navbarSupportedContent"
-          aria-expanded="false"
-          aria-label="Toggle navigation"
-        >
-          <span className="navbar-toggler-icon"></span>
-        </button>
-
-        <div className="collapse navbar-collapse" id="navbarSupportedContent">
-          <ul className="navbar-nav ml-auto">
-            {this.props.items.map((item, idx) => (
-              <li className="nav-item" key={idx} onClick={this.clickHandler}>
-                <a className="nav-link" href={item.link}>
-                  {item.name}
-                </a>
-              </li>
-            ))}
-            {/* <li className="nav-item dropdown">
-              <a
-                className="nav-link dropdown-toggle"
-                href="#"
-                id="navbarDropdown"
-                role="button"
-                data-toggle="dropdown"
-                aria-haspopup="true"
-                aria-expanded="false"
-              >
-                Dropdown
-              </a>
-              <div className="dropdown-menu" aria-labelledby="navbarDropdown">
-                <a className="dropdown-item" href="#">
-                  Action
-                </a>
-                <a className="dropdown-item" href="#">
-                  Another action
-                </a>
-                <div className="dropdown-divider"></div>
-                <a className="dropdown-item" href="#">
-                  Something else here
-                </a>
-              </div>
-            </li> */}
-            {/* <li className="nav-item">
-              <a
-                className="nav-link"
-                href="#"
-                tabindex="-1"
-                aria-disabled="false"
-              >
-                Disabled
-              </a>
-            </li> */}
-          </ul>
-          {/* <form className="form-inline my-2 my-lg-0">
-            <input
-              className="form-control mr-sm-2"
-              type="search"
-              placeholder="Search"
-              aria-label="Search"
-            />
-            <button
-              className="btn btn-outline-success my-2 my-sm-0"
-              type="submit"
-            >
-              Search
-            </button>
-          </form> */}
-        </div>
-      </nav>
-    );
-  }
-}
-
-export default NavBarTest;
